feat: Require a title or decorative prop to be set
Removes the automatic generation of a human readable title, instead requiring either a `title` or the `decorative` prop to be set. This enables automated accessibility testing (e.g. with aXe) without the auto-generated title providing a false positive.

This is important because the icon name (e.g. "Chevron Down") describes the icon and not the function it represents (e.g. "Expand menu options").